% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msissf.R
\name{fit_msissf}
\alias{fit_msissf}
\title{Fit a Markov-Switching integrated Step-Selection Function}
\usage{
fit_msissf(
  formula,
  data,
  N,
  start.values,
  burst_ = "burst_",
  stationary = FALSE,
  iterlim = 1000,
  gradtol = 1e-06,
  decode.states = FALSE,
  print.level = 2,
  stepmax = NULL,
  conf.int = TRUE,
  alpha = 0.05
)
}
\arguments{
\item{formula}{A list with three formulas.}

\item{data}{An object from the \code{amt} package of class \code{random_steps}.}

\item{N}{The number of states.}

\item{start.values}{Starting values, output from \code{generate_starting_values()}.}

\item{burst_}{Column name of columns with the bursts.}

\item{stationary}{logical, TRUE if stationarity for underlying latent MC is assumed}

\item{iterlim}{maximum number of iterations for nlm, default is 1000}

\item{gradtol}{tolerance value for convergence in nlm, default is 1e-06 (default of nlm)}

\item{decode.states}{decode states}

\item{print.level}{print level for nlm, default is 2(all iterations printed)}

\item{stepmax}{maximum allowable scaled step length in nlm, default NULL corresponds to nlm default}

\item{conf.int}{return CI?}

\item{alpha}{alpha value for CI.}
}
\description{
Fit a Markov-Switching integrated Step-Selection Function
}
