% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gernate_starting_values.R
\name{generate_starting_values}
\alias{generate_starting_values}
\title{Generate Starting Values}
\usage{
generate_starting_values(formula, N, data)
}
\arguments{
\item{formula}{A list with three formulas.}

\item{N}{The number of states.}

\item{data}{An object from the \code{amt} package of class \code{random_steps}.}
}
\description{
Helper function to automatically generate starting values given a formula and data.
}
